+++ CHANGELOG.md
@@ -31,7 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for the Fermenter, Mixer, Refinery and Squeezer to Immersive Engineering
 - Added option to ignore creosote output from Coke Oven
 - Added `IngredientHandlers` for 'ItemStack', 'FluidStack', 'Item', 'Block', 'OreDict', 'IIngredient', 'BlockState' and 'SizedObject'.
-- Added `Debug` and `Multithreaded` options to PEIApi config
+- Added `Debug` option to PEIApi config
 
 ### Changed
 - Moved from Travis to GitHub Actions
@@ -65,6 +65,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed PEIApi `LOG` to `LOGGER`
 - Completely changed IngredientHandler internals
 - Cleaned up forEach loop in legacy NuclearCraft plugin
+- Implemented `Callable` interface for APEIPlugin.
 
 ### Removed
 - Harvestcraft jar, we fetch from jitpack now